refactor: consolidate 5 duplicate timeAgo implementations into shared utility
Replace local timeAgo/formatTimeAgo functions in GitHub.jsx, Instances.jsx,
NotificationDropdown.jsx, JobsTab.jsx, and feature-agents/constants.js with
a single timeAgo() exported from utils/formatters.js.
