+++ client/src/components/feature-agents/constants.js
@@ -1,13 +1,5 @@
 import { LayoutDashboard, Settings, Play, GitBranch, Terminal } from 'lucide-react';
-
-export function timeAgo(dateStr) {
-  if (!dateStr) return 'never';
-  const seconds = Math.floor((Date.now() - new Date(dateStr).getTime()) / 1000);
-  if (seconds < 60) return 'just now';
-  if (seconds < 3600) return `${Math.floor(seconds / 60)}m ago`;
-  if (seconds < 86400) return `${Math.floor(seconds / 3600)}h ago`;
-  return `${Math.floor(seconds / 86400)}d ago`;
-}
+export { timeAgo } from '../../utils/formatters';
 
 export const TABS = [
   { id: 'overview', label: 'Overview', icon: LayoutDashboard },
